Directory restructure, final html\css fix.
