+++ src/commands/detectorDiagnostics/models/applensarmresponse.ts
