+++ src/commands/detectorDiagnostics/webviews/networkconnectivity/networkConnectivity.css
@@ -1,32 +1,12 @@
 body {
-	margin: 0;
-	line-height: 1.4285;
-	background-color: #fff;
-	color: #000;
-	font-size: 13px;
-	font-family: "Segoe UI",Segoe,Tahoma,Helvetica,Arial,sans-serif;
-}
-.content-header {
-    font-family: "Segoe UI","Segoe UI Semibold","Segoe WP Semibold","Segoe WP",Tahoma,Arial,sans-serif;
-    font-weight: 600;
-    font-size: 18px;
-    line-height: 30px;
-}
-
-div.nwtable table {
-    border-collapse: collapse;
-    padding: 6px 13px;
-    border: 1px solid black;
+    margin: 0;
+    line-height: 1.4285;
+    background-color: #fff;
+    color: #000;
     font-size: 13px;
+    font-family: "Segoe UI",Segoe,Tahoma,Helvetica,Arial,sans-serif;
 }
 
-div.nwtable tr:nth-child(even) {
-    background-color: #f5f5f5;
-}
-
-div.nwtable {
-    width: 50%;
-}
 .insight-container {
     font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
 }
@@ -37,49 +17,54 @@ div.nwtable {
     border: 1px solid #ddd;
     border-left-width: 0;
     border-radius: 4px;
-    box-shadow: 0 2px 1px rgba(0, 0, 0, 0.05); 
+    box-shadow: 0 2px 1px rgba(0, 0, 0, 0.05);
 }
 
 .insight-container .panel.insight-panel {
     border-color: #ddd;
-    border-left-width: 4px; 	  
+    border-left-width: 4px;
+}
+
+.insight-container .panel.data-container {
+    border: 1px solid #c0d3eb;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
 }
 
 .insight-container .panel.info {
-    border-left-color: #3a9bc7; 
+    border-left-color: #3a9bc7;
 }
 
 .insight-container .panel.info .panel-heading .fa {
-    color: #3a9bc7; 
+    color: #3a9bc7;
 }
 
 .insight-container .panel.success {
-    border-left-color: #3da907; 
+    border-left-color: #3da907;
 }
 
 .insight-container .panel.success .panel-heading .fa {
-    color: #3da907; 
+    color: #3da907;
 }
 
 .insight-container .panel.warning {
-    border-left-color: #ff9104; 
+    border-left-color: #ff9104;
 }
 
 .insight-container .panel.warning .panel-heading .fa {
-    color: #ff9104; 
+    color: #ff9104;
 }
 
 .insight-container .panel.critical {
-    border-left-color: #FF0000; 
+    border-left-color: #FF0000;
 }
 
 .insight-container .panel.critical .panel-heading .fa {
-    color: #FF0000; 
+    color: #FF0000;
 }
 
 .insight-container .panel p {
     margin-top: 0;
-    margin-bottom: .5rem; 
+    margin-bottom: .5rem;
 }
 
 .insight-container .panel .panel-heading {
@@ -91,60 +76,76 @@ div.nwtable {
     color: #333;
     background-color: #f5f5f5;
     border-color: #ddd;
-    font-size: 1.25rem; 
-    cursor: pointer;
+    font-size: 1.25rem;
+}
+
+.insight-container .panel .panel-heading.data-container-heading {
+    background-color: rgba(1, 77, 183, 0.1);
+    color: #012b65;
 }
 
 .insight-container .panel .panel-heading span {
-    display: inline; 
+    display: inline;
 }
 
 .insight-container .panel .panel-heading span.fa {
-    margin-right: 13px; 
+    margin-right: 13px;
 }
 
 .insight-container .panel .panel-heading h5 {
     display: inline;
-    font-family: 'Source Sans Pro',sans-serif;
-    font-weight: normal; 
+    font-family: 'Source Sans Pro', sans-serif;
+    font-weight: normal;
 }
 
 .insight-container .panel .panel-heading h5.panel-title {
     margin-top: 0;
     margin-bottom: 0;
     font-size: 16px;
-    color: inherit; 
+    color: inherit;
 }
 
 .insight-container .panel .panel-heading h5.panel-title .status-icon {
-    margin-right: 25px; 
+    margin-right: 25px;
 }
 
 .insight-container .panel .panel-body {
-    padding: 15px; 
-    font-size: .9rem;
+    padding: 15px;
+}
+
+.insight-container .panel .panel-body.data-container-body table {
+    width: initial !important;
+    margin-bottom: 16px;
+}
+
+.insight-container .panel .panel-body.data-container-body table tr:nth-child(2n) {
+    background-color: rgba(0, 0, 0, 0.03);
+}
+
+.insight-container .panel .panel-body.data-container-body table td, .insight-container .panel .panel-body.data-container-body table th {
+    padding: 6px 13px;
+    border: 1px solid #ddd;
 }
 
 .insight-container .panel .panel-body table {
     width: 100%;
     border-collapse: collapse;
     border-spacing: 0;
-    background-color: transparent; 
+    background-color: transparent;
 }
 
 .insight-container .panel .panel-body table tbody {
     display: table-row-group;
     vertical-align: middle;
-    border-color: inherit; 
+    border-color: inherit;
 }
 
 .insight-container .panel .panel-body table tbody tr td {
     vertical-align: text-top;
-    padding-bottom: 10px; 
+    padding-bottom: 10px;
 }
 
 .insight-container .panel .panel-body table tbody tr td.table-key {
     width: 20%;
-    font-weight: 700; 
+    font-weight: 700;
 }
-
